Replaced render functions into the render.js and moved utils.js into utils folder and renamed utils.js into common.js
