+++ src/main.js
@@ -9,7 +9,7 @@ import TasksComponent from "./components/tasks.js";
 import NoTasksComponent from "./components/no-task.js";
 import {generateTasks} from "./mock/task.js";
 import {generateFilters} from "./mock/filter.js";
-import {render, RenderPosition} from "./utils.js";
+import {render, RenderPosition} from "./utils/render.js";
 
 const TASK_COUNT = 22;
 const SHOWING_TASKS_COUNT_ON_START = 8;
