+++ src/utils/common.js
@@ -1,0 +1,8 @@
+const castTimeFormat = (value) => `0${value}`.slice(-2);
+
+export const formatTime = (date) => {
+  const hours = castTimeFormat(date.getHours() % 12);
+  const minutes = castTimeFormat(date.getMinutes());
+
+  return `${hours}:${minutes}`;
+};
